ros2_ws/src/py_pubsub/requirements.txt
rclpy
pyserial
